Move DataCiteRegistrar changes to existing hack initializer
